+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.2.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- WEB-4041 : Add new "carre" scale
+  [boulch]
 
 
 1.2.5 (2024-01-05)